replace starting numeric prefix with alpha
