+++ packages/cli/openapi-parser/src/v3/utils/getSchemaName.ts
@@ -1,8 +1,13 @@
 import { camelCase, upperFirst } from "lodash-es";
+import { replaceStartingNumber } from "./replaceStartingNumber";
 
 export function getGeneratedTypeName(breadcrumbs: string[]): string {
     const underscoreDelimeted = breadcrumbs.join("_");
-    return upperFirst(camelCase(underscoreDelimeted));
+    const name = upperFirst(camelCase(underscoreDelimeted));
+    if (/^\d/.test(name)) {
+        return replaceStartingNumber(name) ?? name;
+    }
+    return name;
 }
 
 export function getGeneratedPropertyName(breadcrumbs: string[]): string {
